fix: Update download history only when the download button in the download dialog is clicked instead of every time it is opened
